+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Allow adding multiple images to an activity
 - Allow selecting a single video frame as image for an activity
+- The minimum supported android sdk version is now 28
 
 # 2025.11.1